+++ containers/daap-presigned-url/src/var/task/presigned_url.py
@@ -1,14 +1,24 @@
 import json
-import logging
 import os
 import uuid
 from datetime import datetime
 
 import boto3
+from data_platform_logging import DataPlatformLogger
 
-logging.basicConfig(level=logging.INFO, force=True)
-root_logger = logging.getLogger()
 s3 = boto3.client("s3")
+
+logger = DataPlatformLogger(
+    extra={
+        "image_version": os.getenv("VERSION", "unknown"),
+        "base_image_version": os.getenv("BASE_VERSION", "unknown"),
+    }
+)
+
+s3_security_opts = {
+    "ACL": "bucket-owner-full-control",
+    "ServerSideEncryption": "AES256",
+}
 
 
 def handler(event, context):
@@ -58,7 +68,24 @@ def handler(event, context):
         ["content-length-range", 0, 5000000000],
     ]
 
-    root_logger.info(f"s3 key: {file_name}")
+    logger.add_extras(
+        {
+            "lambda_name": context.function_name,
+            "data_product_name": database,
+            "table_name": table,
+        }
+    )
+
+    logger.info(f"s3 bucket: {bucket_name}")
+    logger.info(f"s3 key: {file_name}")
+    logger.info(f"database: {database}")
+    logger.info(f"table: {table}")
+    logger.info(f"amz_date: {amz_date}")
+    logger.info(f"md5: {md5}")
+    logger.info(f"uuid_string: {uuid_string}")
+    logger.info(f"event: {event}")
+
+    logger.write_log_dict_to_s3_json(bucket_name, **s3_security_opts)
 
     # Check the data product has been registered, ie has associated code or metadata in s3
     data_product_registration = s3.list_objects_v2(