+++ containers/daap-presigned-url/CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0]
+
+### Changed
+
+- Use custom logging library image
+- Use custom logging library
+
 ## [1.1.1] 2023-08-30
 
 ### Changed